refactor: centralize runtime command lookup
